Check analog range calibration for change
Instead of detecting simply if the ADC value changed, this now checks if the changed value is outside of the calibrated range (in the same direction it was previously). If so, the user-facing classes aren't going to care and can safely ignore the update.

This was written specifically to solve an issue with the pedals, which don't need to send constant updates when they're at their resting position.