Add deadzone option to pedal serial calibration
This allows the user to set independent 'start' and 'end' deadzones for their pedal serial calibration. The purpose is to prevent pedals from being slightly pressed in their resting position, and to prevent pedals from reaching their 100% "pressed" position at full extension.

By default, this uses a 1% deadzone for the start of travel and a 2.5% deadzone for the end of travel.

I'm trying to be conservative here, because the last thing I want is for a user to set a large deadzone and then be unable to compensate for it when using the pedals as an HID device. Still, I think a tiny bit of fuzzing is good and can compensate for noise at the input extremes.