Add activeLow option to peripherals
Although all current peripherals use active-high detection, that may not be the case in the future. Changing this now for future compatibility.

The Logitech classes maintain their original arguments, as we know they are active-high.

Because the added argument includes a default, this does not technically break the public API.