Swap G25/G27 LED and detect pin order
This is arbitrary, but it's important to get this change in there before the API is locked for the major release.

Originally, I figured that resistors were required for both pins (pull-down for the detect, series for the LED). In which case, it's more likely that users would want the detect feature (disabling inputs on disconnect) over the LED feature (...blinky).

After some testing I've determined that even though the LED should probably have a 100 Ohm resistor, it will survive without it. That means it's more likely for users to set it, in which case it should take precedence over the detect pin.

In addition, the LED pin is shared by the data input of the EEPROM chip. Although the library does not currently interface with the EEPROM, if it ever does that pin will become required. In which case, again, the detect pin would end up at the end of the list so it can have a default argument.